+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add custom config values via `C::Config()->set(...)` which are stored in the AppStorage for runtime only
   (but can also be stored in AppStorage cache depending on user's app)
 - Model's `filterBasedOnRequest()` can now also filter fields with custom callback, check for `isnull` / `notnull`
+- `C::Formatter()->removeTrailingZeros(...)` method which removes trailing zeros and dots
 
 ### Changed
 - Improved security of DebugBar